move files to same package, update test args
